Updated Tenjin iOS SDK to 1.5.1 + Updated AIR to 23.0.0.257

+++ ios/TenjinIosExtension/TenjinIosExtension/FRETypeConverter.m
@@ -76,7 +76,7 @@
         return FRE_INVALID_ARGUMENT;
     }
     const char* utf8String = string.UTF8String;
-    unsigned long length = strlen( utf8String );
+    uint32_t length = (uint32_t) strlen( utf8String );
     return FRENewObjectFromUTF8( length + 1, (uint8_t*) utf8String, asObject );
 }
